+++ app/lib/shared/versions.dart
@@ -24,8 +24,9 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// when the version switch happens.
 const _acceptedRuntimeVersions = <String>[
   // The current [runtimeVersion].
-  '2026.05.04',
+  '2026.05.05',
   // Fallback runtime versions.
+  '2026.05.04',
   '2026.04.29',
 ];
 
@@ -52,10 +53,7 @@ String get runtimeVersion => acceptedRuntimeVersions.first;
 /// The version which marks the earliest version of the data which we'd like to
 /// keep during various GC processes. Data prior to this version is subject to
 /// delete (unless there is another rule in place to keep it).
-String get gcBeforeRuntimeVersion => acceptedRuntimeVersions.length > 1
-    ? acceptedRuntimeVersions.last
-    // TODO: remove this after the new indexed blob is deployed
-    : '2026.04.07';
+String get gcBeforeRuntimeVersion => acceptedRuntimeVersions.last;
 
 /// Returns true if the given version should be considered as obsolete and can
 /// be deleted.
@@ -65,7 +63,7 @@ bool shouldGCVersion(String version) =>
 // keep in-sync with SDK version in .github/workflows/all-test.yml and Dockerfile.app
 final String runtimeSdkVersion = '3.11.0';
 final String toolStableDartSdkVersion = '3.11.5';
-final String toolStableFlutterSdkVersion = '3.41.7';
+final String toolStableFlutterSdkVersion = '3.41.9';
 
 final semanticToolStableDartSdkVersion = Version.parse(
   toolStableDartSdkVersion,